Merging my fork with the master repository (#1)
* MagicBird Version 1.0

These are all of the files needed to run magic bird correctly.

The main program is testTemplate.py

That program will open up the template.txt file and read in the information given. From there it will utilize the decode.py program to search through all of the json files in the JSON Files folder and write any tweets that meet the given criteria (keywords, geotag, etc) and write the data from those tweets into csv files and store them in the TempDir folder. With those csv files, the testTemplate.py program will now utilize the combine.py program to take the files in the TempDir folder and combine them by month or all and then store the new csv file(s) in the Final Form folder.

For the current template.txt file it will get all tweets with the words @EastLosHighShow, check, and Picture (while not worrying about geotag) and then combine them by month which is just July (07) and August (08).

* Delete decoder.py

* Delete decoderClass.py

* Delete csv_combine.py

* Delete emo_and_class.py