Merging... (#4)
* MagicBird Version 1.0

These are all of the files needed to run magic bird correctly.

The main program is testTemplate.py

That program will open up the template.txt file and read in the information given. From there it will utilize the decode.py program to search through all of the json files in the JSON Files folder and write any tweets that meet the given criteria (keywords, geotag, etc) and write the data from those tweets into csv files and store them in the TempDir folder. With those csv files, the testTemplate.py program will now utilize the combine.py program to take the files in the TempDir folder and combine them by month or all and then store the new csv file(s) in the Final Form folder.

For the current template.txt file it will get all tweets with the words @EastLosHighShow, check, and Picture (while not worrying about geotag) and then combine them by month which is just July (07) and August (08).

* Delete decoder.py

* Delete decoderClass.py

* Delete csv_combine.py

* Delete emo_and_class.py

* Bug Fix + New Mentioned Users Added into CSV files

A bug fix that changed the value of the truncated field to a boolean variable

New feature added to append all mentioned users in a tweet to the very end of the tweet data in the CSV files
     Mentioned users are put at the end because there could be more than one, and each screen name will be contained in its own cell

* Updated: Version 2.1

Contains functionality to process and format the new Tweet format provided by Twitter

* Bug Fixes

* Delete README.txt

* bug fixes

* Add files via upload

* Add files via upload

* Add files via upload

* Delete testTemplate.py

* Delete template.txt

* Delete emojilist4.csv

* Delete decode.py

* Delete combine.py

* Delete README.txt

* Delete combine.cpython-36.pyc

* Delete decode.cpython-36.pyc

* Delete 20170422000000_data.csv

* Delete 20160802000909_data.csv

* Delete 20160801000245_data.csv

* Delete 20160714000615_data.csv

* Delete 20160711165412_data.csv

* Delete 20160711165412_ELH_stream.json

* Delete 20160714000615_ELH_stream.json

* Delete 20160801000245_ELH_stream.json

* Delete 20160802000909_ELH_stream.json

* Delete 04MasterCSV.csv

* Delete 07MasterCSV.csv

* Delete 08MasterCSV.csv

* New Name, Same Game

Just updated the name and file paths to RITHM

* Update template.txt

* Update testTemplate.py

* Delete combine.py

* Delete decode.py

* Delete emojilist4.csv

* Delete template.txt

* Delete testTemplate.py