Adjusted handling of "extended_tweet" field.
Handles of original posters do not appear in the "extended_tweet/full_text" field in the case of retweets. This caused unintended behavior of skipping over retweets, when searching for the twitter handle of the original poster. For example, if @ColditzJB sends a tweet, and @MTH_Pitt retweets it, that retweet will now appear in results when searching for "@ColditzJB" in the parser. This allows the tracking of message rebroadcasting related to for a Twitter user of interest. 

Future updates should make this an optional behavior, so that these can be ignored if desired.