Incorporated ts() and mkdir() functions
"ts()" makes datetime formatting consistent and callable across the code base. It will output current time and also convert Twitter timestamps to "YYYYMMDDhhmmss" format. "mkdir()" is a helper function for creating new directories when they are needed (to check for and create a directory before trying to write new files to it).